+++ src/test-helpers/fixtures.ts
@@ -47,17 +47,17 @@ export const biomeInstalledPackageJson = {
 // Sample biome.json data
 export const baseBiomeConfig = {
 	$schema: "https://biomejs.dev/schemas/1.9.4/schema.json",
-	extends: ["@mfyuu/biome-config/base"],
+	extends: ["@lemtoc/biome-config/base"],
 };
 
 export const reactBiomeConfig = {
 	$schema: "https://biomejs.dev/schemas/1.9.4/schema.json",
-	extends: ["@mfyuu/biome-config/react"],
+	extends: ["@lemtoc/biome-config/react"],
 };
 
 export const nextBiomeConfig = {
 	$schema: "https://biomejs.dev/schemas/1.9.4/schema.json",
-	extends: ["@mfyuu/biome-config/next"],
+	extends: ["@lemtoc/biome-config/next"],
 };
 
 // Sample VS Code settings.json data